# pastel360-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PASTEL 360º API RESTFUL",
        "description": "API para gerenciamento de clientes, produtos e pedidos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Lista todos os clientes",
                "operationId": "d00d20521584ab5da83e088abb7aded2",
                "responses": {
                    "200": {
                        "description": "Lista de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Cria um novo cliente",
                "operationId": "f1e86e5b79fedfa6593067977292b13b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "mail",
                                    "phone",
                                    "birthdate",
                                    "place",
                                    "number",
                                    "zipcode",
                                    "district"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "José Romano"
                                    },
                                    "mail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jromano@teste.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1996-01-01"
                                    },
                                    "place": {
                                        "type": "string",
                                        "example": "Rua Principal"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "zipcode": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "complement": {
                                        "type": "string",
                                        "example": "Apto 45"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Obtém um cliente específico",
                "operationId": "4c3c7071ea1133f04bacdbc70f8aa230",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Atualiza um cliente",
                "operationId": "7d0a5ae9e306858d30e04fc50265d80c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "José Romano"
                                    },
                                    "mail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jromano@teste.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1996-01-01"
                                    },
                                    "place": {
                                        "type": "string",
                                        "example": "Rua Principal"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "zipcode": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "complement": {
                                        "type": "string",
                                        "example": "Apto 45"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Exclui um cliente",
                "operationId": "c929f14f421a80a896c6fe116ffdc317",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente excluído com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Client": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "José Romano"
                    },
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "jromano@teste.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "11999999999"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "example": "1996-01-01"
                    },
                    "place": {
                        "type": "string",
                        "example": "Rua Principal"
                    },
                    "number": {
                        "type": "string",
                        "example": "123"
                    },
                    "zipcode": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "district": {
                        "type": "string",
                        "example": "Centro"
                    },
                    "complement": {
                        "type": "string",
                        "example": "Apto 45"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}